LayoutTests/imported/w3c/web-platform-tests/css/selectors/child-indexed-pseudo-class-expected.txt


FAIL Expected HTML element to match :first-child with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :last-child with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :only-child with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :first-of-type with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :last-of-type with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :only-of-type with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :nth-child(1) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :nth-child(n) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :nth-last-child(1) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :nth-last-child(n) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :nth-of-type(1) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :nth-of-type(n) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :nth-last-of-type(1) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected HTML element to match :nth-last-of-type(n) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
PASS Expected HTML element to not match :nth-child(2) with matches, querySelector(), and querySelectorAll()
PASS Expected HTML element to not match :nth-last-child(2) with matches, querySelector(), and querySelectorAll()
PASS Expected HTML element to not match :nth-of-type(2) with matches, querySelector(), and querySelectorAll()
PASS Expected HTML element to not match :nth-last-of-type(2) with matches, querySelector(), and querySelectorAll()
FAIL Expected DIV element to match :first-child with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :last-child with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :only-child with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :first-of-type with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :last-of-type with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :only-of-type with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-child(1) with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-child(n) with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-last-child(1) with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-last-child(n) with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-of-type(1) with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-of-type(n) with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-last-of-type(1) with matches assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-last-of-type(n) with matches assert_equals: expected false but got true
PASS Expected DIV element to not match :nth-child(2) with matches
PASS Expected DIV element to not match :nth-last-child(2) with matches
PASS Expected DIV element to not match :nth-of-type(2) with matches
PASS Expected DIV element to not match :nth-last-of-type(2) with matches
FAIL Expected DIV element to match :first-child with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :last-child with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :only-child with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :first-of-type with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :last-of-type with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :only-of-type with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-child(1) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-child(n) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-last-child(1) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-last-child(n) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-of-type(1) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-of-type(n) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-last-of-type(1) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
FAIL Expected DIV element to match :nth-last-of-type(n) with matches, querySelector(), and querySelectorAll() assert_equals: expected false but got true
PASS Expected DIV element to not match :nth-child(2) with matches, querySelector(), and querySelectorAll()
PASS Expected DIV element to not match :nth-last-child(2) with matches, querySelector(), and querySelectorAll()
PASS Expected DIV element to not match :nth-of-type(2) with matches, querySelector(), and querySelectorAll()
PASS Expected DIV element to not match :nth-last-of-type(2) with matches, querySelector(), and querySelectorAll()
